Update controller tests to use shared examples, etc
