+++ spec/controllers/hyc/downloads_controller_spec.rb
@@ -5,10 +5,10 @@ RSpec.describe Hyc::DownloadsController, type: :controller do
   describe '#track_download' do
     before do
       Hyrax.config.google_analytics_id = 'blah'
+      stub_request(:post, "http://www.google-analytics.com/collect").to_return(status: 200, body: "", headers: {})
     end
 
     it "has the method for tracking analytics for download" do
-      stub_request(:post, "http://www.google-analytics.com/collect").to_return(status: 200, body: "", headers: {})
       expect(controller).to respond_to(:track_download)
       expect(controller.track_download).to be_a_kind_of Net::HTTPOK
     end
@@ -19,14 +19,24 @@ RSpec.describe Hyc::DownloadsController, type: :controller do
     let(:solr_response) { {response: {docs: [{admin_set_tesim: ['admin set for download controller']}]}}.to_json }
     let(:empty_solr_response) { {response: {docs: []}}.to_json }
 
-    it 'finds admin set for file set' do
-      stub_request(:get, /solr/).to_return(body: solr_response)
-      expect(controller.set_record_admin_set).to eq('admin set for download controller')
+    context 'with a solr response' do
+      before do
+        stub_request(:get, /solr/).to_return(body: solr_response)
+      end
+
+      it 'finds admin set for file set' do
+        expect(controller.set_record_admin_set).to eq('admin set for download controller')
+      end
     end
 
-    it 'does not find admin set for file set' do
-      stub_request(:get, /solr/).to_return(body: empty_solr_response)
-      expect(controller.set_record_admin_set).to eq('Unknown')
+    context 'with an empty solr response' do
+      before do
+        stub_request(:get, /solr/).to_return(body: empty_solr_response)
+      end
+
+      it 'does not find admin set for file set' do
+        expect(controller.set_record_admin_set).to eq('Unknown')
+      end
     end
   end
 end
