+++ spec/controllers/omniauth_controller_spec.rb
@@ -10,15 +10,19 @@ RSpec.describe OmniauthController, type: :request do
     end
 
     context 'not using database auth' do
-      it 'is successful' do
-        cached_database_auth = ENV['DATABASE_AUTH']
+      before do
+        @cached_database_auth = ENV['DATABASE_AUTH']
         ENV['DATABASE_AUTH'] = 'false'
         allow(Rails).to receive(:env).and_return(ActiveSupport::StringInquirer.new("production"))
+      end
 
+      after do
+        ENV['DATABASE_AUTH'] = @cached_database_auth
+      end
+
+      it 'is successful' do
         get new_user_session_path
         expect(response).to redirect_to "/Shibboleth.sso/Login?target=#{user_shibboleth_omniauth_authorize_path}%26origin%3D"
-
-        ENV['DATABASE_AUTH'] = cached_database_auth
       end
     end
   end
@@ -33,18 +37,22 @@ RSpec.describe OmniauthController, type: :request do
     end
 
     context 'not using database auth' do
-      it 'redirects to shibboleth logout url' do
-        cached_database_auth = ENV['DATABASE_AUTH']
-        cached_sso_logout_url = ENV['SSO_LOGOUT_URL']
+      before do
+        @cached_database_auth = ENV['DATABASE_AUTH']
+        @cached_sso_logout_url = ENV['SSO_LOGOUT_URL']
         ENV['DATABASE_AUTH'] = 'false'
         ENV['SSO_LOGOUT_URL'] = 'https://shibboleth.example.com/idp/logout.jsp'
         allow(Rails).to receive(:env).and_return(ActiveSupport::StringInquirer.new("production"))
+      end
 
+      after do
+        ENV['DATABASE_AUTH'] = @cached_database_auth
+        ENV['SSO_LOGOUT_URL'] = @cached_sso_logout_url
+      end
+
+      it 'redirects to shibboleth logout url' do
         get destroy_user_session_path
         expect(response).to redirect_to 'https://shibboleth.example.com/idp/logout.jsp'
-
-        ENV['DATABASE_AUTH'] = cached_database_auth
-        ENV['SSO_LOGOUT_URL'] = cached_sso_logout_url
       end
     end
   end
